requirements.txt
fastapi
uvicorn 
pymongo 
pydantic 
gspread 
oauth2client 
faker 
google-genai
python-dotenv
